requirements.txt
streamlit
pymongo
python-dotenv
bcrypt
streamlit-authenticator
pandas
plotly